@robodoo 2024-05-25
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/166784).