@zaha-odoo 2024-05-25
Memory benchmarks before/after commit
[memray.zip](https://github.com/odoo/odoo/files/15441028/memray.zip)