@moylop260 2024-05-28
@zaha-odoo 

What about adding unittest with assert `self.assertQueryCount(expected_query_count)`?