@zaha-odoo 2024-05-28
@moylop260 
Specifically for `_rollup_move_xxx_fetch` ?