@zaha-odoo 2024-05-29
> @zaha-odoo
> 
> What about adding unittest with assert `self.assertQueryCount(expected_query_count)`?

This is a good idea. I may try this soon if it proves not too difficult. It would be assuring to have, and the logistics team reviewer(s) would appreciate it.