@zaha-odoo 2024-05-29
> Call fetch to get the state and quantity fields from all sibling moves. These could instead be added into the consolidated prefetch_ids. Maybe that is more preferable, but I think fetch here is OK because we only need those two fields from the sibling moves.

Regarding this in my original comment, the sibling moves I refer to are these: https://github.com/odoo/odoo/blob/60989c7da8b1811ff0c106ce85109e857e72e96c/addons/stock/report/stock_forecasted.py#L229-L231