@zaha-odoo 2024-06-03
Closing in favor of https://github.com/odoo/odoo/pull/167806
Due to this PR's branch name being bad.